xenon upd

+++ XenonResurrection/SimulatorXenon2/CicParserSim/cpu.h
@@ -77,8 +77,8 @@ extern reg_t _reg;
 #define zf _reg.flags.bit.zero
 
 
-WORD& memory16(WORD segment, WORD offset);
-BYTE& memory(WORD segment, WORD offset);
+WORD& memory16(WORD segment, int offset);
+BYTE& memory(WORD segment, int offset);
 
 void _push(WORD w);
 WORD _pop();
@@ -114,18 +114,18 @@ void _in(BYTE& value, WORD port);
 
 struct MemVideo
 {
-    static BYTE Get8(WORD seg, WORD nAddr);
-    static void Set8(WORD seg, WORD nAddr, BYTE nData);
-    static WORD Get16(WORD seg, WORD nAddr);
-    static void Set16(WORD seg, WORD nAddr, WORD nData);
+    static BYTE Get8(WORD seg, int nAddr);
+    static void Set8(WORD seg, int nAddr, BYTE nData);
+    static WORD Get16(WORD seg, int nAddr);
+    static void Set16(WORD seg, int nAddr, WORD nData);
 };
 
 struct MemData
 {
-    static BYTE Get8(WORD seg, WORD nAddr);
-    static void Set8(WORD seg, WORD nAddr, BYTE nData);
-    static WORD Get16(WORD seg, WORD nAddr);
-    static void Set16(WORD seg, WORD nAddr, WORD nData);
+    static BYTE Get8(WORD seg, int nAddr);
+    static void Set8(WORD seg, int nAddr, BYTE nData);
+    static WORD Get16(WORD seg, int nAddr);
+    static void Set16(WORD seg, int nAddr, WORD nData);
 };
 
 
@@ -421,8 +421,8 @@ void _repne_scasb()
     }
 }
 
-BYTE memoryVideoGet(WORD seg, WORD ofs);
-void memoryVideoSet(WORD seg, WORD ofs, BYTE data);
+BYTE memoryVideoGet(WORD seg, int ofs);
+void memoryVideoSet(WORD seg, int ofs, BYTE data);
 void _sync();
 void _idiv(WORD d)
 {
